Moved facebook ads to gss operator to google.cloud #8502
