+++ docs/autoapi_templates/index.rst
@@ -110,8 +110,6 @@ All operators are in the following packages:
 
   airflow/providers/exasol/operators/index
 
-  airflow/providers/google/facebook_ads_to_gcs/operators/index
-
   airflow/providers/ftp/sensors/index
 
   airflow/providers/google/ads/operators/index
