+++ docs/operators-and-hooks-ref.rst
@@ -802,7 +802,7 @@ These integrations allow you to copy data from/to Google Cloud Platform.
    * - `Facebook Ads <http://business.facebook.com>`__
      - `Google Cloud Storage (GCS) <https://cloud.google.com/gcs/>`__
      -
-     - :mod:`airflow.providers.google.facebook_ads_to_gcs.operators.ads`
+     - :mod:`airflow.providers.google.cloud.operators.facebook_ads_to_gcs`
 
    * - `Google BigQuery <https://cloud.google.com/bigquery/>`__
      - `MySQL <https://www.mysql.com/>`__
